refactor: changed the local to implement api-key management functions
