+++ base-cli/cmd/common/auth/types.go
@@ -78,70 +78,8 @@ type TokenExchangeResult struct {
 	Scope        []string  `json:"scope"`
 }
 
-// ApiKeys representa o retorno da requisição de API Keys
-type ApiKeys struct {
-	UUID          string  `json:"uuid"`
-	Name          string  `json:"name"`
-	ApiKey        string  `json:"api_key"`
-	Description   string  `json:"description"`
-	KeyPairID     string  `json:"key_pair_id"`
-	KeyPairSecret string  `json:"key_pair_secret"`
-	StartValidity string  `json:"start_validity"`
-	EndValidity   *string `json:"end_validity,omitempty"`
-	RevokedAt     *string `json:"revoked_at,omitempty"`
-	TenantName    *string `json:"tenant_name,omitempty"`
-	Tenant        struct {
-		UUID      string `json:"uuid"`
-		LegalName string `json:"legal_name"`
-	} `json:"tenant"`
-	Scopes []struct {
-		UUID        string `json:"uuid"`
-		Name        string `json:"name"`
-		Title       string `json:"title"`
-		ConsentText string `json:"consent_text"`
-		Icon        string `json:"icon"`
-		APIProduct  struct {
-			UUID string `json:"uuid"`
-			Name string `json:"name"`
-		} `json:"api_product"`
-	} `json:"scopes"`
-}
-
 type Scope struct {
 	Name  string `json:"name"`
 	Title string `json:"title"`
 	UUID  string `json:"uuid"`
-}
-
-type APIProduct struct {
-	Name   string  `json:"name"`
-	Scopes []Scope `json:"scopes"`
-	UUID   string  `json:"uuid"`
-}
-
-type Platform struct {
-	APIProducts []APIProduct `json:"api_products"`
-	Name        string       `json:"name"`
-	UUID        string       `json:"uuid"`
-}
-
-type PlatformsResponse []Platform
-
-type ScopesCreate struct {
-	ID            string `json:"id"`
-	RequestReason string `json:"request_reason"`
-}
-
-type CreateApiKey struct {
-	Name          string         `json:"name"`
-	Description   string         `json:"description"`
-	TenantID      string         `json:"tenant_id"`
-	ScopesList    []ScopesCreate `json:"scopes"`
-	StartValidity string         `json:"start_validity"`
-	EndValidity   string         `json:"end_validity"`
-}
-
-type ApiKeyResult struct {
-	UUID string `json:"uuid,omitempty"`
-	Used bool   `json:"used,omitempty"`
 }
